Refactor attachment delegate and actions to a separate class.
