+++ Example/Example/UIImage+SaveTo.swift
@@ -1,0 +1,21 @@
+import Foundation
+import UIKit
+
+extension UIImage {
+
+    func saveToTemporaryFile() -> URL {
+        let fileName = "\(ProcessInfo.processInfo.globallyUniqueString)_file.jpg"
+
+        guard let data = self.jpegData(compressionQuality: 0.9) else {
+            fatalError("Could not conert image to JPEG.")
+        }
+
+        let fileURL = URL(fileURLWithPath: NSTemporaryDirectory()).appendingPathComponent(fileName)
+
+        guard (try? data.write(to: fileURL, options: [.atomic])) != nil else {
+            fatalError("Could not write the image to disk.")
+        }
+
+        return fileURL
+    }
+}
